pyproject.toml
[tool.poetry]
name = "showcase"
version = "0.1.0"
description = ""
authors = ["Samuel Baetz <samuelbaetz105@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
Pillow = "^8.1.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
